Added further explanation for "Unknown" perms
Added detection of a permission object that has multiple entries. In my testing this has occurred only when there is both a Shared Drive AND a file permission in place and the Shared Drive permission is automatically overriding the file permission. The shared drive permission takes array position 0 and the file permission takes position 1 in this case. 

Minor formatting cleanup as well.