ALPHA CODE / 1st Commit
This has been used in production to remove a user from every file / folder shared to them within a shared drive. 

Note - there are many limitations to this as it is now... proceed with EXTREME caution. Do not use --do-it until you really understand what this is doing. Also note that this will require full read / write Google Drive scope access.Whatever user runs this script will appear to "edit" the files / folders where a permission removal is executed.